Did assignment 16

+++ 16_subseq/maxSeq.c
@@ -2,9 +2,9 @@
 #include<stdlib.h>
 
 size_t maxSeq(int * array, size_t n)
-{size_t count=1;
+{int count=1;
   size_t temp=0;
-  size_t i;
+  int i;
   if(n==0)
     {return 0;
     }
